+++ src/buganize/__main__.py
@@ -1,0 +1,3 @@
+from buganize.cli.main import start
+
+start()